Add files via upload
Evader
========

It's an exe packer (and a Crypter) which will encrypt your PE exe input file and add it as resource to the end of the output new exe file The encryption key size and complexity can be given as input

Usage of packer (Encryption)
packer.exe <input-path> <output-path> <key-size> <start-ascii> <end-ascii>
The complexity of encryption key will be determined by and

for example this command will lead to keys from AAAA to ZZZZ

packer.exe <input-path> <output-path> 4 65 90
The Decryption part
Decryption isn't like other packers stub and it's based on BruteForcing and examining each character to retrieve the encryption key! why!?

And after retrieving key the encrypted payload will be decrypted and will be run directly inside memory and in form of char[] which is contained each byte of the main PE file!